+++ includes/Tuple.hpp
@@ -37,7 +37,7 @@ struct tuple
     }
 
     // Getter: float f = t[0];
-    float operator[](int index) const {
+    float operator[](size_t index) const {
         switch (index) {
             case 0: return x;
             case 1: return y;
@@ -48,7 +48,7 @@ struct tuple
     }
 
     // Setter: t[0] = 5.0f;
-    float& operator[](int index) {
+    float& operator[](size_t index) {
         switch (index) {
             case 0: return x;
             case 1: return y;
